Added Contributing Guidelines and Link in Readme
Introduced a new file, CONTRIBUTING.md, to provide guidelines for potential contributors. This includes an introduction to our CI/CD pipeline, steps on how to join the project, and standards for commit messages and issue creation. Additionally, updated README.md by adding a reference to the new CONTRIBUTING.md file for better discoverability of contribution instructions.