Subject: Removed Git dependencies and added ImGui library files
In this update, we removed the dependency to clone ImGui from GitHub in the "DevDeps.sh" script. We made this decision to simplify the process and avoid any potential issues that could arise due to changes in the ImGui repository, and to minimize network requirements during installation. Instead, we added the necessary ImGui library files directly to our repository, thereby making the project self-contained and ensuring consistent setups. Consequently, please note that it is necessary to commit ImGui updates into this repository in order for them to be reflected in the project.