chore: Add Cypress for web-interface tests
This commit adds Cypress as a development dependency for running tests on the web-interface. The Cypress package is installed using npm, and the necessary scripts for running Cypress tests are added to the package.json file. This will enable running Cypress tests as part of the test workflow for the web-interface.